Update rplidar_scan_publisher.cpp
Fixing a compile issue on Ubuntu 22.04 running ROS2 Humble.  
Upon compilation of a fresh 'git clone', colcon build failed to compile the package, identifying missing type's from the "this->declare_parameter" functions in the initialisation section of the node.  To resolve this and build, I've combined the declarations into one section.